new functions and breakout publications utils
